[Feature] Combine Web Cmdlet Partial Class Files
