+++ src/Microsoft.PowerShell.Commands.Utility/commands/utility/WebCmdlet/Common/BasicHtmlWebResponseObject.Common.cs
@@ -10,6 +10,7 @@ using System.Text;
 using System.Text.RegularExpressions;
 using System.Collections.Generic;
 using System.Diagnostics;
+using System.Net.Http;
 
 namespace Microsoft.PowerShell.Commands
 {
@@ -240,6 +241,50 @@ namespace Microsoft.PowerShell.Commands
             {
                 this.Content = string.Empty;
             }
+        }
+
+        #endregion Methods
+    }
+
+    // TODO: Merge Partials
+
+    // <summary>
+    /// Response object for html content without DOM parsing
+    /// </summary>
+    public partial class BasicHtmlWebResponseObject : WebResponseObject
+    {
+        #region Constructors
+
+        /// <summary>
+        /// Constructor for BasicHtmlWebResponseObject
+        /// </summary>
+        /// <param name="response"></param>
+        public BasicHtmlWebResponseObject(HttpResponseMessage response)
+            : this(response, null)
+        { }
+
+        /// <summary>
+        /// Constructor for HtmlWebResponseObject with memory stream
+        /// </summary>
+        /// <param name="response"></param>
+        /// <param name="contentStream"></param>
+        public BasicHtmlWebResponseObject(HttpResponseMessage response, Stream contentStream)
+            : base(response, contentStream)
+        {
+            EnsureHtmlParser();
+            InitializeContent();
+            InitializeRawContent(response);
+        }
+
+        #endregion Constructors
+
+        #region Methods
+
+        private void InitializeRawContent(HttpResponseMessage baseResponse)
+        {
+            StringBuilder raw = ContentHelper.GetRawContentHeader(baseResponse);
+            raw.Append(Content);
+            this.RawContent = raw.ToString();
         }
 
         #endregion Methods
